feat: Refactor register component
